refactor: 패키지 domain, view 설정
패키지 domain 생성, UI → view 변경
Car, RacingCar domain으로 이동
패키지명 변경에 따른 변경사항 전체적용
