+++ src/docs/3단계 - 자동차 경주.md
@@ -56,9 +56,9 @@ int number = scanner.nextInt();
 
 ## 프로그래밍 요구사항
 
-- [x] 모든 로직에 단위 테스트를 구현한다. 단, UI(System.out, System.in) 로직은 제외
+- [x] 모든 로직에 단위 테스트를 구현한다. 단, Car.view(System.out, System.in) 로직은 제외
   - 핵심 로직을 구현하는 코드와 UI를 담당하는 로직을 구분한다.
-  - UI 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
+  - Car.view 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
 - [x] 자바 코드 컨벤션을 지키면서 프로그래밍한다.
   - 참고문서: https://google.github.io/styleguide/javaguide.html 또는 https://myeonguni.tistory.com/1596
 - [x] else 예약어를 쓰지 않는다.
