+++ src/main/java/Car/domain/Car.java
@@ -1,4 +1,4 @@
-package Car;
+package Car.domain;
 
 
 import java.util.Random;
