+++ src/test/java/Car/CarTest.java
@@ -1,5 +1,6 @@
 package Car;
 
+import Car.domain.Car;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
