+++ src/test/java/Car/RacingCarTest.java
@@ -1,7 +1,8 @@
 package Car;
 
+import Car.domain.Car;
+import Car.domain.RacingCar;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
